feat(drive): replace import page with overlay and fix destination to current folder
